+++ main/ver_06/Customer.cpp
@@ -3,7 +3,7 @@
 Customer::Customer(){
 }
 
-Customer::Customer(string name, string address, string phone, string from_date, string to_date, double advance_payment, int booking_id)
+Customer::Customer(string name, string address, string phone, string from_date, string to_date, long double advance_payment, int booking_id)
     : name(name), address(address), phone(phone), from_date(from_date), to_date(to_date), advance_payment(advance_payment), booking_id(booking_id) {
 }
 
@@ -25,7 +25,7 @@ void Customer::setFromDate(string from_date){
 void Customer::setToDate(string to_date){
     this->to_date = to_date;
 }
-void Customer::setAdvancePayment(double advance_payment){
+void Customer::setAdvancePayment(long double advance_payment){
     this->advance_payment = advance_payment;
 }
 void Customer::setBookingId(int booking_id){
@@ -47,7 +47,7 @@ string Customer::getFromDate() const{
 string Customer::getToDate() const{
     return this->to_date;
 }
-double Customer::getAdvancePayment() const{
+long double Customer::getAdvancePayment() const{
     return this->advance_payment;
 }
 int Customer::getBookingId() const{
