+++ main/ver_06/Customer.h
@@ -11,18 +11,18 @@ private:
     string phone;
     string from_date;
     string to_date;
-    double advance_payment;
+    long double advance_payment;
     int booking_id;
 public:
     Customer();
-    Customer(string,string,string,string,string,double,int);
+    Customer(string,string,string,string,string,long double,int);
     ~Customer();
     void setName(string);
     void setAddress(string);
     void setPhone(string);
     void setFromDate(string);
     void setToDate(string);
-    void setAdvancePayment(double);
+    void setAdvancePayment(long double);
     void setBookingId(int);
 
     string getName() const;
@@ -30,6 +30,6 @@ public:
     string getPhone() const;
     string getFromDate() const;
     string getToDate() const;
-    double getAdvancePayment() const;
+    long double getAdvancePayment() const;
     int getBookingId() const;
 };
