+++ main/ver_06/Room.h
@@ -6,23 +6,23 @@ class Room
 private:
     bool status;
     int roomNumber;
-    double rent;
+    long double rent;
     bool type;
     // Customer customer;
 
 public:
     Room();
-    Room(bool, int, double, bool);
+    Room(bool, int, long double, bool);
     ~Room();
     void setType(bool);
     void setRoomNumber(int);
-    void setRent(double);
+    void setRent(long double);
     void setStatus(bool);
     // void setCustomer(Customer &);
 
     bool getType() const;
     int getRoomNumber() const;
-    double getRent() const;
+    long double getRent() const;
     bool getStatus() const;
     // Customer getCustomer() const;
